// frontend/generated/models/StockKeepingUnitDto.ts
export interface StockKeepingUnitDto {
    /**
     * 
     * @type {number}
     * @memberof StockKeepingUnitDto
     */
    skuId?: number;
    /**
     * 
     * @type {string}
     * @memberof StockKeepingUnitDto
     */
    name?: string;
    /**
     * 
     * @type {ImageDto}
     * @memberof StockKeepingUnitDto
     */
    image?: ImageDto;
    /**
     * 
     * @type {number}
     * @memberof StockKeepingUnitDto
     */
    price?: number;
    /**
     * 
     * @type {number}
     * @memberof StockKeepingUnitDto
     */
    stock?: number;
    /**
     * 
     * @type {number}
     * @memberof StockKeepingUnitDto
     */
    specCount?: number;
    /**
     * 
     * @type {BlindBoxDto}
     * @memberof StockKeepingUnitDto
     */
    blindBox?: BlindBoxDto;
    /**
     * 
     * @type {Date}
     * @memberof StockKeepingUnitDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof StockKeepingUnitDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof StockKeepingUnitDto
     */
    isVisible?: boolean;
}